+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.26.2
+- [`ready`] automatically correct non-standard chaid id (ex. A-2 to B)
+- [`ready`] set `--separate-hetgen` as False by default
+- [`ready`] fixer saves fixed receptor instead of fixed complex
+- [`ready`] handles presence of multiple ligands (ex. two ligands in each protomer in a dimer)
+- [`relax`] charge assignment for multiple ligands
+   
 ## 0.17.0
 - added `--desmond` option to `mdworks equi` command
 - added `relax` protocol